Moved cancellation related code to separate file. Consist parameter names.
