+++ packages/vscode-messenger-common/src/index.ts
@@ -5,3 +5,4 @@
  ******************************************************************************/
 
 export * from './messages';
+export * from './cancellation';
